use sublime exec in ride menu

+++ ride/main_menu.py
@@ -1,6 +1,7 @@
 import sublime
 import sublime_plugin
 import os
+from shutil import copyfile
 import threading
 
 _main_menu_is_visible = [False]
@@ -76,8 +77,10 @@ class RideMainMenuListener(sublime_plugin.EventListener):
 
         def set_main_menu():
 
-            menu_path = os.path.join(sublime.packages_path(), 'User', 'RIDE',
-                                     'Main.sublime-menu')
+            menu_path = os.path.join(
+                sublime.packages_path(), 'User', 'RIDE', 'Main.sublime-menu')
+            user_menu_path = os.path.join(
+                sublime.packages_path(), 'User', 'RIDE', 'RIDE.sublime-menu')
             menu_dir = os.path.dirname(menu_path)
 
             if self.is_r_project(view.window()) or self.is_r_file(view):
@@ -86,11 +89,14 @@ class RideMainMenuListener(sublime_plugin.EventListener):
                     os.makedirs(menu_dir, 0o755)
 
                 if not os.path.exists(menu_path):
-                    data = sublime.load_resource(
-                        "Packages/RIDE/support/RIDE.sublime-menu")
-                    with open(menu_path, 'w') as f:
-                        f.write(data.replace("\r\n", "\n"))
-                        f.close()
+                    if os.path.exists(user_menu_path):
+                        copyfile(user_menu_path, menu_path)
+                    else:
+                        data = sublime.load_resource(
+                            "Packages/RIDE/support/RIDE.sublime-menu")
+                        with open(menu_path, 'w') as f:
+                            f.write(data.replace("\r\n", "\n"))
+                            f.close()
                 _main_menu_is_visible[0] = True
             else:
                 if os.path.exists(menu_path):
@@ -101,9 +107,41 @@ class RideMainMenuListener(sublime_plugin.EventListener):
         self.timer.start()
 
 
-class RidePackageSendCodeCommand(sublime_plugin.WindowCommand):
+class RidePackageExecCommand(sublime_plugin.WindowCommand):
     def is_visible(self):
         return self.window.id() in _window_is_rproject
 
     def run(self, **kwargs):
-        self.window.active_view().run_command("send_code", kwargs)
+        kwargs["working_dir"] = self.window.folders()[0]
+        self.window.run_command("exec", kwargs)
+
+
+class RideRenderRmarkdownCommand(sublime_plugin.TextCommand):
+    def is_enabled(self):
+        return self.view.settings().get("syntax").endswith("R Markdown.sublime-syntax")
+
+    def run(self, edit):
+        cmd = "rmarkdown::render(\"$file\", encoding = \"UTF-8\")"
+        self.view.run_command("send_code", {"cmd": cmd})
+
+
+class RideSweaveRnwCommand(sublime_plugin.TextCommand):
+    def is_enabled(self):
+        return self.view.settings().get("syntax").endswith("R Sweave.sublime-syntax")
+
+    def run(self, edit):
+        cmd = ("""setwd(\"$file_path\")\n"""
+               """Sweave(\"$file\")\n"""
+               """tools::texi2dvi(\"$file_base_name.tex\", pdf = TRUE)""")
+        self.view.run_command("send_code", {"cmd": cmd})
+
+
+class RideKnitRnwCommand(sublime_plugin.TextCommand):
+    def is_enabled(self):
+        return self.view.settings().get("syntax").endswith("R Sweave.sublime-syntax")
+
+    def run(self, edit):
+        cmd = ("""setwd(\"$file_path\")\n"""
+               """knitr::knit(\"$file\", output=\"$file_base_name.tex\")\n"""
+               """tools::texi2dvi(\"$file_base_name.tex\", pdf = TRUE)")""")
+        self.view.run_command("send_code", {"cmd": cmd})
